use UnlimitedLookupCache from master in test
