+++ src/test/java/com/fasterxml/jackson/databind/type/ContainerTypesTest.java
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.BaseMapTest;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.util.LRUMap;
+import com.fasterxml.jackson.databind.util.UnlimitedLookupCache;
 
 // for [databind#1415]
 public class ContainerTypesTest extends BaseMapTest
@@ -51,7 +52,7 @@ public class ContainerTypesTest extends BaseMapTest
 
     public void testCustomLookupCache() throws Exception
     {
-        TypeFactory tf = MAPPER.getTypeFactory().withCache(new ConcurrentMapLookupCache<Object,JavaType>());
+        TypeFactory tf = MAPPER.getTypeFactory().withCache(new UnlimitedLookupCache<Object, JavaType>(0));
         JavaType t = tf.constructParametricType(List.class, HashMap.class);
         assertEquals(CollectionType.class, t.getClass());
         assertEquals(List.class, t.getRawClass());
